refactor: move template backend test classes to test_utils
